+++ src/server/scenario_parser/WormDTO.cpp
@@ -3,7 +3,7 @@
 
 using std::string;
 
-WormDTO::WormDTO(int id, string image, int x, int y): id(id), image(image), x(x), y(y) {}
+WormDTO::WormDTO(int id, string image, float x, float y): id(id), image(image), x(x), y(y) {}
 
 int WormDTO::get_id() {
     return id;
@@ -13,10 +13,10 @@ string WormDTO::get_image() {
     return image;
 }
 
-int WormDTO::get_x() {
+float WormDTO::get_x() {
     return x;
 }
 
-int WormDTO::get_y() {
+float WormDTO::get_y() {
     return y;
 }
